♻️  refactor: 무한스크롤 코드 변경 (#28)
