+++ src/main/kotlin/com/kotlin/study/dongambackend/domain/comment/entity/Comment.kt
@@ -18,15 +18,15 @@ class Comment : BaseTimeEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.SEQUENCE)
     @Column(name = "id", nullable = false)
-    var id: Long? = null
+    val id: Long? = null
 
     // TODO: userId 참조 필요
     @Column(name = "user_id", nullable = false)
-    var userId: Long? = 1
+    var userId: Long = 1
 
     // TODO: postId 참조 필요
     @Column(name = "post_id", nullable = false)
-    var postId: Long? = 1
+    var postId: Long = 1
 
     var content: String? = null
 
@@ -38,8 +38,8 @@ class Comment : BaseTimeEntity {
         content = _content
     }
 
-    fun updateComment(commnetUpdateRequest: CommentUpdateRequest) {
-        content = commnetUpdateRequest.content
+    fun updateComment(commentUpdateRequest: CommentUpdateRequest) {
+        content = commentUpdateRequest.content
     }
 
     fun deleteComment() {
